requirements.txt
bleak>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
urllib3>=2.0.0
